+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 - Ability to get DSNPAddDelegate logs
 - Ability to get DSNPRemoveDelegate logs
 - createdAt field to GraphChange announcement type, plus update to parquet schema.
+- Ability to get all identities associated to an address
 ### Removed
 -
 ### Fixed